ya se pueden ver los viajes de manera general. Estoy avanzando con la creacion de gastos y ver el resultado final.

+++ app/src/main/java/com/anwera64/pagodividido/domain/models/Companion.kt
@@ -1,3 +1,3 @@
 package com.anwera64.pagodividido.domain.models
 
-class Companion(val uid: String, val name: String, val totalDebt: Int, val totalOwed: Int)
+class Companion(val uid: String, val name: String, val totalDebt: Float, val totalOwed: Float)
